LayoutTests/media/track/track-css-stroke-cues-expected.txt


Test that style to all cues is applied correctly.
EVENT(canplaythrough)
EVENT(seeked)
EXPECTED (getComputedStyle(textTrackDisplayElement(video, 'cue')).webkitTextStrokeWidth == '4px') OK
EXPECTED (getComputedStyle(textTrackDisplayElement(video, 'cue')).webkitTextStrokeColor == 'rgb(0, 0, 0)') OK
EXPECTED (getComputedStyle(textTrackDisplayElement(video, 'cue')).paintOrder == 'stroke') OK
EXPECTED (getComputedStyle(textTrackDisplayElement(video, 'cue')).strokeLinejoin == 'round') OK
EXPECTED (getComputedStyle(textTrackDisplayElement(video, 'cue')).strokeLinecap == 'round') OK
EXPECTED (getComputedStyle(textTrackDisplayElement(video, 'cue')).overflow == 'visible') OK
END OF TEST
